Allow negative longitude angles when regridding from LatLon grids
Previously, the longitudes were assumed to be in the 0...360 range, and
working with data where longitudes were in the -180...180 range meant
first converting the angles then rearranging the data such that the
longitudes were in ascending order.

When building the regridder, we now check the source longitudes for any
negative values, and if any are found, then we map the destination
longitudes to the -180...180 range.

This is a format common to some datasets (e.g. ERA5) and makes it easier
to work with such data without having to do preprocessing steps.